requirements.txt
fastapi=0.81.0
mysql-connector-python=8.3.0
uvicorn=0.18.3
python-multipart=0.0.5
pyjwt=2.3.0